requirements.txt
numpy
streamlit==1.3.1
tensorflow-cpu==2.8.0
opencv-python-headless==4.5.1.48
streamlit-webrtc==0.11.0 